Make private ValidationAttributeAdapter
