+++ Sharprompt/Internal/PropertyMetadata.cs
@@ -6,6 +6,8 @@ using System.Reflection;
 
 using Sharprompt.Forms;
 
+using AnnotationsDataType = System.ComponentModel.DataAnnotations.DataType;
+
 namespace Sharprompt.Internal
 {
     internal class PropertyMetadata
@@ -16,21 +18,21 @@ namespace Sharprompt.Internal
             var dataTypeAttribute = propertyInfo.GetCustomAttribute<DataTypeAttribute>();
 
             PropertyInfo = propertyInfo;
-            PropertyType = Nullable.GetUnderlyingType(propertyInfo.PropertyType) ?? propertyInfo.PropertyType;
-            DataType = dataTypeAttribute?.DataType;
+            Type = Nullable.GetUnderlyingType(propertyInfo.PropertyType) ?? propertyInfo.PropertyType;
             IsCollection = propertyInfo.PropertyType.IsGenericType && propertyInfo.PropertyType.GetGenericTypeDefinition() == typeof(IEnumerable<>);
+            DataType = dataTypeAttribute?.DataType;
             Message = displayAttribute?.GetPrompt();
             Order = displayAttribute?.GetOrder();
             DefaultValue = propertyInfo.GetValue(model);
             Validators = propertyInfo.GetCustomAttributes<ValidationAttribute>(true)
-                                     .Select(x => new ValidationAttributeAdapter(x).GetValidator(PropertyInfo.Name, model))
+                                     .Select(x => new ValidationAttributeAdapter(x).GetValidator(propertyInfo.Name, model))
                                      .ToArray();
         }
 
         public PropertyInfo PropertyInfo { get; }
-        public Type PropertyType { get; }
-        public DataType? DataType { get; }
+        public Type Type { get; }
         public bool IsCollection { get; }
+        public AnnotationsDataType? DataType { get; }
         public string Message { get; }
         public int? Order { get; }
         public object DefaultValue { get; }
@@ -38,22 +40,22 @@ namespace Sharprompt.Internal
 
         public FormType DetermineFormType()
         {
-            if (DataType == System.ComponentModel.DataAnnotations.DataType.Password)
+            if (DataType == AnnotationsDataType.Password)
             {
                 return FormType.Password;
             }
 
-            if (PropertyType == typeof(bool))
+            if (Type == typeof(bool))
             {
                 return FormType.Confirm;
             }
 
-            if (PropertyType.IsEnum)
+            if (Type.IsEnum)
             {
                 return FormType.Select;
             }
 
-            if (IsCollection && PropertyType.GetGenericArguments()[0].IsEnum)
+            if (IsCollection && Type.GetGenericArguments()[0].IsEnum)
             {
                 return FormType.MultiSelect;
             }
@@ -64,6 +66,27 @@ namespace Sharprompt.Internal
             }
 
             return FormType.Input;
+        }
+
+        private class ValidationAttributeAdapter
+        {
+            public ValidationAttributeAdapter(ValidationAttribute validationAttribute)
+            {
+                _validationAttribute = validationAttribute;
+            }
+
+            private readonly ValidationAttribute _validationAttribute;
+
+            public Func<object, ValidationResult> GetValidator(string propertyName, object model)
+            {
+                var validationContext = new ValidationContext(model)
+                {
+                    DisplayName = propertyName,
+                    MemberName = propertyName
+                };
+
+                return input => _validationAttribute.GetValidationResult(input, validationContext);
+            }
         }
     }
 }
